+++ lib/domain/record/record_page_store.dart
@@ -10,7 +10,7 @@ import 'package:pilll/entity/pill_sheet_type.dart';
 import 'package:pilll/entity/setting.codegen.dart';
 import 'package:pilll/native/legacy.dart';
 import 'package:pilll/service/auth.dart';
-import 'package:pilll/service/pill_sheet.dart';
+import 'package:pilll/database/pill_sheet.dart';
 import 'package:pilll/domain/record/record_page_state.codegen.dart';
 import 'package:pilll/database/pill_sheet_group.dart';
 import 'package:pilll/database/pill_sheet_modified_history.dart';
