+++ lib/domain/settings/setting_page_store.dart
@@ -4,7 +4,7 @@ import 'dart:io';
 import 'package:pilll/database/batch.dart';
 import 'package:pilll/entity/setting.codegen.dart';
 import 'package:pilll/native/health_care.dart';
-import 'package:pilll/service/pill_sheet.dart';
+import 'package:pilll/database/pill_sheet.dart';
 import 'package:pilll/database/pill_sheet_group.dart';
 import 'package:pilll/database/pill_sheet_modified_history.dart';
 import 'package:pilll/service/setting.dart';
